Drop Ruby 2.4 and make CI simpler
Configuring GitHub to have many required tests is hard to maintain.

In fact, the admin needs to duplicate what is defined in the actions.yml test matrix. It should be quite a bit easier this way, just having one CI job that depends on the success of the entire test matrix.